+++ CHANGES.md
@@ -1,3 +1,13 @@
+## HJN-1.0.3
+#### ADDED
+- node focus
+- node movement
+- node creation
+- contextmenu node search
+## FIXED
+- link limitations
+- node unique ids
+
 ## HJN-1.0.2
 #### ADDED
 - node links